+++ DEV_ONLY/index.ts
@@ -1,5 +1,5 @@
 import Bluebird from 'bluebird';
-import { deepEqual, shallowEqual } from 'fast-equals';
+import { deepEqual } from 'fast-equals';
 
 import memoize from '../src';
 import { __ } from 'ramda';
@@ -73,31 +73,37 @@ console.groupEnd();
 console.group('maxArgs');
 
 // limit to testing the first args
-const isEqualMaxArgs = (
-  originalKey: MicroMemoize.Key,
-  newKey: MicroMemoize.Key,
+const isMatchingKeyMaxArgs = (
+  originalKey: string[],
+  newKey: string[],
 ): boolean => {
   return originalKey[0] === newKey[0];
 };
 
-const memoizedMax = memoize(method, { isEqual: isEqualMaxArgs });
+const memoizedMax = memoize(method, { isMatchingKey: isMatchingKeyMaxArgs });
 
 memoizedMax(foo, bar);
-memoizedMax(foo, 'baz');
-memoizedMax(foo, 'quz');
+memoizedMax(foo, baz);
+memoizedMax(foo, quz);
 
 console.groupEnd();
 
 console.group('custom - deep equals');
 
-const deepEqualMethod = ({ one, two }: { one: string; two: string }) => {
+const deepEqualMethod = ({
+  one,
+  two,
+}: {
+  one: string | number;
+  two: string | number;
+}) => {
   console.log('custom equal method fired', one, two);
 
   return [one, two];
 };
 
 const deepEqualMemoized = memoize(deepEqualMethod, {
-  isEqual: deepEqual,
+  isMatchingKey: deepEqual,
 });
 
 deepEqualMemoized({ one: 1, two: 2 });
@@ -182,6 +188,7 @@ const withDefault = (foo: string, bar: string = 'default') => {
 
   return `${foo} ${bar}`;
 };
+'';
 const moizedWithDefault = memoize(withDefault, { maxSize: 5 });
 
 console.log(moizedWithDefault(foo));
@@ -199,10 +206,10 @@ const noFns = (one: string, two: string, three: Function) => {
 };
 
 const memoizedNoFns = memoize(noFns, {
-  isEqual(key1: MicroMemoize.Key, key2: MicroMemoize.Key) {
+  isEqual(key1: string, key2: string) {
     return key1 === key2;
   },
-  transformKey(args: MicroMemoize.RawKey) {
+  transformKey(args: any) {
     return [JSON.stringify(args)];
   },
 });
@@ -237,11 +244,16 @@ console.log(matchingKeyMemoized.cache);
 
 console.groupEnd();
 
+type Dictionary<Type> = {
+  [key: string]: Type;
+  [index: number]: Type;
+};
+
 const calc = memoize(
-  (object: PlainObject, metadata: PlainObject): PlainObject =>
-    Object.keys(object).reduce((totals: PlainObject, key: string) => {
+  (object: Dictionary<any>, metadata: Dictionary<any>) =>
+    Object.keys(object).reduce((totals: Dictionary<any>, key: string) => {
       if (Array.isArray(object[key])) {
-        totals[key] = object[key].map((subObject: PlainObject) =>
+        totals[key] = object[key].map((subObject: Dictionary<any>) =>
           calc(subObject, metadata),
         );
       } else {